moved rest of auto to auto_commands folder
