+++ src/main/java/frc/robot/commands/auto_commands/AutoMid.java
@@ -5,22 +5,20 @@
 /* the project.                                                               */
 /*----------------------------------------------------------------------------*/
 
-package frc.robot.commands;
+package frc.robot.commands.auto_commands;
 
-import edu.wpi.first.wpilibj2.command.InstantCommand;
-import frc.robot.Robot;
+import edu.wpi.first.wpilibj2.command.SequentialCommandGroup;
 
 // NOTE:  Consider using this command inline, rather than writing a subclass.  For more
 // information, see:
 // https://docs.wpilib.org/en/latest/docs/software/commandbased/convenience-features.html
-public class AutoBackward extends InstantCommand {
-  public AutoBackward() {
-    // Use addRequirements() here to declare subsystem dependencies.
-  }
-
-  // Called when the command is initially scheduled.
-  @Override
-  public void initialize() {
-    Robot.driveSubsystem.driveSystem(-.5, 0);
+public class AutoMid extends SequentialCommandGroup {
+  /**
+   * Creates a new AutoMid.
+   */
+  public AutoMid() {
+    // Add your commands in the super() call, e.g.
+    // super(new FooCommand(), new BarCommand());
+    addCommands(new AutoShoot().withTimeout(5), new AutoBackward().withTimeout(1));
   }
 }
